fixes
reparo de muchas cosas

+++ Services/Actividades/ActividadesCAService.cs
@@ -52,7 +52,7 @@ namespace ControlActividades.Services.Actividades
                         Descripcion = a.Descripcion,
                         FechaCreacion = a.FechaCreacion,
                         FechaLimite = a.FechaLimite,
-                        Puntaje = a.Puntaje
+                        Puntaje = (int)a.Puntaje
                     })
                     .ToListAsync();
 
@@ -89,7 +89,7 @@ namespace ControlActividades.Services.Actividades
                     MateriaId = entidad.MateriaId,
                     FechaCreacion = entidad.FechaCreacion,
                     FechaLimite = entidad.FechaLimite,
-                    Puntaje = entidad.Puntaje,
+                    Puntaje = (int)entidad.Puntaje,
                     Enviado = entidad.Enviado,
                     PermitirEntregasTarde = entidad.PermitirEntregasTarde,
                     FechaProgramada = entidad.FechaProgramada
@@ -145,7 +145,7 @@ namespace ControlActividades.Services.Actividades
                 Descripcion = actividadEditar.Descripcion,
                 FechaCreacion = actividadEditar.FechaCreacion,
                 FechaLimite = actividadEditar.FechaLimite,
-                Puntaje = actividadEditar.Puntaje
+                Puntaje = (int)actividadEditar.Puntaje
             };
             
         }
